fix: alinha tipo das FKs (int -> long) com Entity.Id
Entity.Id (classe base) e long, mas todas as propriedades de FK (OficinaId, ClienteId, VeiculoId, PecaId, FornecedorId, OrdemServicoId, TipoServicoId, ServicoId, FuncionarioResponsavelId) estavam declaradas como int, causando incompatibilidade de tipo entre chave primaria e chave estrangeira. Alterado para long em: Fornecedor, Funcionario, MovimentacaoEstoque, Orcamento, OrdemServico, OrdemServicoServico, Peca, Servico, TipoServico e Veiculo.

+++ src/GRA.Domain/Entities/Fornecedor.cs
@@ -4,7 +4,7 @@ namespace GRA.Domain.Entities;
 
 public class Fornecedor : Entity
 {
-    public int OficinaId { get; set; }
+    public long OficinaId { get; set; }
     public Oficina? Oficina { get; set; }
 
     public required string RazaoSocial { get; set; }
