+++ src/GRA.Domain/Entities/OrdemServicoServico.cs
@@ -4,13 +4,13 @@ namespace GRA.Domain.Entities;
 
 public class OrdemServicoServico : Entity
 {
-    public int OficinaId { get; set; }
+    public long OficinaId { get; set; }
     public Oficina? Oficina { get; set; }
 
-    public int OrdemServicoId { get; set; }
+    public long OrdemServicoId { get; set; }
     public OrdemServico? OrdemServico { get; set; }
 
-    public int ServicoId { get; set; }
+    public long ServicoId { get; set; }
     public Servico? Servico { get; set; }
 
     public string? Observacoes { get; set; }
